components/model_evaluation/native-requirements.txt
components==1.2.8
joblib==1.1.0
numpy==1.23.2
pandas==1.4.3
python_dateutil==2.8.2
scikit_learn==1.1.2
scipy==1.9.1
SQLAlchemy==1.4.40
utils==1.0.1
joblib==1.1.0
scikit_learn==1.1.2
